qemu test working with manual copy of linux files

+++ flowey/flowey_lib_hvlite/src/init_vmm_tests_content_dir.rs
@@ -19,6 +19,7 @@ use crate::build_tmks::TmksOutput;
 use crate::build_tpm_guest_tests::TpmGuestTestsOutput;
 use crate::build_vmgstool::VmgstoolOutput;
 use crate::common::CommonArch;
+use crate::common::CommonTriple;
 use crate::download_release_igvm_files_from_gh::OpenhclReleaseVersion;
 use flowey::node::prelude::*;
 
@@ -37,6 +38,11 @@ macro_rules! define_vmm_tests_built_artifacts {
             pub struct VmmTestsBuiltArtifactsWrite {
                 $(pub $artifact: Option<::flowey::node::prelude::WriteVar<$output>>,)*
             }
+
+            #[derive(Serialize, Deserialize, Default, Debug)]
+            pub struct VmmTestsBuiltArtifactsSelections {
+                $(pub $artifact: bool,)*
+            }
         }
     };
 }
@@ -52,7 +58,11 @@ define_vmm_tests_built_artifacts!(
     openvmm => OpenvmmOutput,
     openvmm_vhost => OpenvmmVhostOutput,
     pipette_windows => PipetteOutput,
-    pipette_linux_musl => PipetteOutput,
+    // Specify arch for pipette here as a hack to bring up qemu support
+    // TODO: have a VmmTestsBuiltArtifacts for each target with corresponding
+    // test content sub-dir.
+    pipette_linux_musl_x64 => PipetteOutput,
+    pipette_linux_musl_aarch64 => PipetteOutput,
     guest_test_uefi => GuestTestUefiOutput,
     openhcl_standard => OpenhclIgvmOutput,
     openhcl_standard_dev => OpenhclIgvmOutput,
@@ -122,6 +132,17 @@ macro_rules! vmm_tests_built_artifacts_builder {
     };
 }
 
+#[derive(Serialize, Deserialize, Debug, Default)]
+pub struct VmmTestsPreBuiltArtifactsSelections {
+    pub test_linux_initrd: bool,
+    pub test_linux_kernel: bool,
+    pub test_linux_bzimage: bool,
+    pub uefi: bool,
+    pub virtio_win_drivers: bool,
+    pub release_igvm: bool,
+    pub qemu_system_aarch64: bool,
+}
+
 flowey_request! {
     pub struct Request {
         /// Directory to symlink / copy test contents into. Does not need to be
@@ -134,12 +155,12 @@ flowey_request! {
         pub vmm_tests_target: target_lexicon::Triple,
         /// Artifacts used by the tests that are built in the openvmm repo
         pub built_artifacts: VmmTestsBuiltArtifacts,
+        /// Artifacts to download that are pre-built as part of OpenVMM deps
+        pub prebuilt_artifacts: VmmTestsPreBuiltArtifactsSelections,
         /// Copy files necessary to use the test content dir as a minimal repo root.
         ///
         /// This is useful for running tests on machines without a local clone.
         pub is_repo_root: bool,
-        /// Whether the tests require that release igvm files are downloaded
-        pub needs_release_igvm: bool,
         /// Whether to copy incubator profiles into the test content directory.
         pub needs_incubator_profiles: bool,
 
@@ -159,6 +180,7 @@ impl SimpleFlowNode for Node {
         ctx.import::<crate::git_checkout_openvmm_repo::Node>();
         ctx.import::<crate::download_uefi_mu_msvm::Node>();
         ctx.import::<crate::download_release_igvm_files_from_gh::resolve::Node>();
+        ctx.import::<crate::resolve_openvmm_qemu::Node>();
     }
 
     fn process_request(request: Self::Request, ctx: &mut NodeCtx<'_>) -> anyhow::Result<()> {
@@ -166,8 +188,8 @@ impl SimpleFlowNode for Node {
             test_content_dir,
             vmm_tests_target,
             built_artifacts,
+            prebuilt_artifacts,
             is_repo_root,
-            needs_release_igvm,
             needs_incubator_profiles,
             done,
         } = request;
@@ -177,46 +199,57 @@ impl SimpleFlowNode for Node {
 
         let arch = CommonArch::from_architecture(vmm_tests_target.architecture)?;
 
-        let test_linux_initrd =
-            ctx.reqv(|v| crate::resolve_openvmm_test_initrd::Request::Get(arch, v));
-        let test_linux_kernel = ctx.reqv(|v| {
-            crate::resolve_openvmm_test_linux_kernel::Request::Get(
-                crate::resolve_openvmm_test_linux_kernel::OpenvmmTestKernelFile::Kernel,
-                arch,
-                crate::resolve_openvmm_test_linux_kernel::DEFAULT_LINUX_TEST_KERNEL_VERSION,
-                v,
-            )
+        let test_linux_initrd = prebuilt_artifacts
+            .test_linux_initrd
+            .then(|| ctx.reqv(|v| crate::resolve_openvmm_test_initrd::Request::Get(arch, v)));
+        let test_linux_kernel = prebuilt_artifacts.test_linux_kernel.then(|| {
+            ctx.reqv(|v| {
+                crate::resolve_openvmm_test_linux_kernel::Request::Get(
+                    crate::resolve_openvmm_test_linux_kernel::OpenvmmTestKernelFile::Kernel,
+                    arch,
+                    crate::resolve_openvmm_test_linux_kernel::DEFAULT_LINUX_TEST_KERNEL_VERSION,
+                    v,
+                )
+            })
         });
-        let test_linux_bzimage =
-            crate::resolve_openvmm_test_linux_kernel::OpenvmmTestKernelFile::BzImage
-                .is_available_for(arch)
-                .then(|| {
-                    ctx.reqv(|v| {
-                        crate::resolve_openvmm_test_linux_kernel::Request::Get(
-                            crate::resolve_openvmm_test_linux_kernel::OpenvmmTestKernelFile::BzImage,
-                            arch,
-                            crate::resolve_openvmm_test_linux_kernel::DEFAULT_LINUX_TEST_KERNEL_VERSION,
-                            v,
-                        )
-                    })
-                });
+        let test_linux_bzimage = prebuilt_artifacts.test_linux_bzimage.then(|| {
+            ctx.reqv(|v| {
+                crate::resolve_openvmm_test_linux_kernel::Request::Get(
+                    crate::resolve_openvmm_test_linux_kernel::OpenvmmTestKernelFile::BzImage,
+                    arch,
+                    crate::resolve_openvmm_test_linux_kernel::DEFAULT_LINUX_TEST_KERNEL_VERSION,
+                    v,
+                )
+            })
+        });
 
-        let uefi =
-            ctx.reqv(|v| crate::download_uefi_mu_msvm::Request::GetMsvmFd { arch, msvm_fd: v });
+        let uefi = prebuilt_artifacts.uefi.then(|| {
+            ctx.reqv(|v| crate::download_uefi_mu_msvm::Request::GetMsvmFd { arch, msvm_fd: v })
+        });
 
-        let virtio_win_dir = ctx.reqv(crate::resolve_openvmm_test_virtio_win::Request::Get);
+        let virtio_win_dir = prebuilt_artifacts
+            .virtio_win_drivers
+            .then(|| ctx.reqv(crate::resolve_openvmm_test_virtio_win::Request::Get));
 
-        let release_igvm_files = if needs_release_igvm {
-            Some(ctx.reqv(
+        let release_igvm_files = prebuilt_artifacts.release_igvm.then(|| {
+            ctx.reqv(
                 |v| crate::download_release_igvm_files_from_gh::resolve::Request {
                     arch,
                     release_igvm_files: v,
                     release_version: OpenhclReleaseVersion::latest(),
                 },
-            ))
-        } else {
-            None
-        };
+            )
+        });
+
+        let qemu_system_aarch64 = prebuilt_artifacts.qemu_system_aarch64.then(|| {
+            ctx.reqv(|v| {
+                crate::resolve_openvmm_qemu::Request::Get(
+                    crate::resolve_openvmm_qemu::QemuFile::SystemAarch64,
+                    arch,
+                    v,
+                )
+            })
+        });
 
         let VmmTestsBuiltArtifacts {
             flowey_hvlite,
@@ -227,7 +260,8 @@ impl SimpleFlowNode for Node {
             openvmm,
             openvmm_vhost,
             pipette_windows,
-            pipette_linux_musl,
+            pipette_linux_musl_x64,
+            pipette_linux_musl_aarch64,
             guest_test_uefi,
             openhcl_standard,
             openhcl_standard_dev,
@@ -268,7 +302,8 @@ impl SimpleFlowNode for Node {
                     openvmm,
                     openvmm_vhost,
                     pipette_windows,
-                    pipette_linux_musl,
+                    pipette_linux_musl_x64,
+                    pipette_linux_musl_aarch64,
                     guest_test_uefi,
                     openhcl_igvm_files,
                     tmks,
@@ -285,6 +320,7 @@ impl SimpleFlowNode for Node {
                     uefi,
                     virtio_win_dir,
                     release_igvm_files,
+                    qemu_system_aarch64,
                 )
             );
 
@@ -296,6 +332,7 @@ impl SimpleFlowNode for Node {
                 let test_linux_bzimage = test_linux_bzimage.map(|v| rt.read(v));
                 let uefi = rt.read(uefi);
                 let release_igvm_files_dir = rt.read(release_igvm_files);
+                let qemu_system_aarch64 = rt.read(qemu_system_aarch64);
                 let test_content_dir = rt.read(test_content_dir);
 
                 if !test_content_dir.exists() {
@@ -404,17 +441,29 @@ impl SimpleFlowNode for Node {
                     }
                 }
 
-                if let Some(pipette_linux) = pipette_linux_musl {
-                    match rt.read(pipette_linux) {
-                        PipetteOutput::LinuxBin { bin, dbg: _ } => {
-                            let dst = test_content_dir.join("pipette");
-                            fs_err::copy(bin, &dst)?;
-                            dst.make_executable()?;
-                        }
-                        _ => {
-                            anyhow::bail!("did not find `pipette` in RegisterPipetteLinuxMusl")
-                        }
+                let mut write_pipette_linux = |target: CommonTriple, pipette| match rt.read(pipette)
+                {
+                    PipetteOutput::LinuxBin { bin, dbg: _ } => {
+                        let target_dir = test_content_dir.join(target.to_string());
+                        let dst = target_dir.join("pipette");
+                        fs_err::create_dir_all(&target_dir)?;
+                        fs_err::copy(&bin, &dst)?;
+                        dst.make_executable()?;
+                        Ok(())
                     }
+                    _ => {
+                        anyhow::bail!("did not find `pipette` in RegisterPipetteLinuxMusl")
+                    }
+                };
+
+                if let Some(pipette_linux_musl_x64) = pipette_linux_musl_x64 {
+                    write_pipette_linux(CommonTriple::X86_64_LINUX_MUSL, pipette_linux_musl_x64)?;
+                }
+                if let Some(pipette_linux_musl_aarch64) = pipette_linux_musl_aarch64 {
+                    write_pipette_linux(
+                        CommonTriple::AARCH64_LINUX_MUSL,
+                        pipette_linux_musl_aarch64,
+                    )?;
                 }
 
                 if let Some(guest_test_uefi) = guest_test_uefi {
@@ -554,19 +603,30 @@ impl SimpleFlowNode for Node {
                     }
                 }
 
+                if let Some(qemu_system_aarch64) = qemu_system_aarch64 {
+                    fs_err::copy(
+                        qemu_system_aarch64,
+                        test_content_dir.join("qemu-system-aarch64"),
+                    )?;
+                }
+
                 let (arch_dir, kernel_file_name) = match arch {
                     CommonArch::X86_64 => ("x64", "vmlinux"),
                     CommonArch::Aarch64 => ("aarch64", "Image"),
                 };
                 fs_err::create_dir_all(test_content_dir.join(arch_dir))?;
-                fs_err::copy(
-                    test_linux_initrd,
-                    test_content_dir.join(arch_dir).join("initrd"),
-                )?;
-                fs_err::copy(
-                    test_linux_kernel,
-                    test_content_dir.join(arch_dir).join(kernel_file_name),
-                )?;
+                if let Some(test_linux_initrd) = test_linux_initrd {
+                    fs_err::copy(
+                        test_linux_initrd,
+                        test_content_dir.join(arch_dir).join("initrd"),
+                    )?;
+                }
+                if let Some(test_linux_kernel) = test_linux_kernel {
+                    fs_err::copy(
+                        test_linux_kernel,
+                        test_content_dir.join(arch_dir).join(kernel_file_name),
+                    )?;
+                }
                 if let Some(bzimage_path) = test_linux_bzimage {
                     fs_err::copy(
                         bzimage_path,
@@ -574,18 +634,20 @@ impl SimpleFlowNode for Node {
                     )?;
                 }
 
-                let uefi_dir = test_content_dir.join(match arch {
-                    CommonArch::Aarch64 => {
-                        "hyperv.uefi.mscoreuefi.AARCH64.RELEASE/MsvmAARCH64/RELEASE_CLANGPDB/FV"
-                    }
-                    CommonArch::X86_64 => {
-                        "hyperv.uefi.mscoreuefi.x64.RELEASE/MsvmX64/RELEASE_VS2022/FV"
-                    }
-                });
-                fs_err::create_dir_all(&uefi_dir)?;
-                fs_err::copy(uefi, uefi_dir.join("MSVM.fd"))?;
+                if let Some(uefi) = uefi {
+                    let uefi_dir = test_content_dir.join(match arch {
+                        CommonArch::Aarch64 => {
+                            "hyperv.uefi.mscoreuefi.AARCH64.RELEASE/MsvmAARCH64/RELEASE_CLANGPDB/FV"
+                        }
+                        CommonArch::X86_64 => {
+                            "hyperv.uefi.mscoreuefi.x64.RELEASE/MsvmX64/RELEASE_VS2022/FV"
+                        }
+                    });
+                    fs_err::create_dir_all(&uefi_dir)?;
+                    fs_err::copy(uefi, uefi_dir.join("MSVM.fd"))?;
+                }
 
-                {
+                if let Some(virtio_win_dir) = virtio_win_dir {
                     let src = rt.read(virtio_win_dir);
                     let dst = test_content_dir.join("virtio-win");
                     let _ = fs_err::remove_dir_all(&dst);
@@ -624,7 +686,7 @@ pub mod vmm_tests_artifact_builders {
             nextest_vmm_tests_archive => NextestVmmTestsArchive,
             openvmm => OpenvmmOutput,
             openvmm_vhost => OpenvmmVhostOutput,
-            pipette_linux_musl => PipetteOutput,
+            pipette_linux_musl_x64 => PipetteOutput,
             prep_steps => PrepStepsOutput,
             // any machine
             guest_test_uefi => GuestTestUefiOutput,
@@ -650,7 +712,7 @@ pub mod vmm_tests_artifact_builders {
             openhcl_standard => OpenhclIgvmOutput,
             openhcl_cvm => OpenhclIgvmOutput,
             openhcl_linux_direct => OpenhclIgvmOutput,
-            pipette_linux_musl => PipetteOutput,
+            pipette_linux_musl_x64 => PipetteOutput,
             tmk_vmm_linux_musl => TmkVmmOutput,
             // any machine
             guest_test_uefi => GuestTestUefiOutput,
@@ -670,7 +732,7 @@ pub mod vmm_tests_artifact_builders {
             vmgstool_dev => VmgstoolOutput,
             // linux build machine
             openhcl_standard => OpenhclIgvmOutput,
-            pipette_linux_musl => PipetteOutput,
+            pipette_linux_musl_aarch64 => PipetteOutput,
             tmk_vmm_linux_musl => TmkVmmOutput,
             // any machine
             guest_test_uefi => GuestTestUefiOutput,
@@ -690,7 +752,7 @@ pub mod vmm_tests_artifact_builders {
             // aarch64 guest binaries
             nextest_vmm_tests_archive => NextestVmmTestsArchive,
             openvmm => OpenvmmOutput,
-            pipette_linux_musl => PipetteOutput,
+            pipette_linux_musl_aarch64 => PipetteOutput,
             guest_test_uefi => GuestTestUefiOutput,
             tmks => TmksOutput,
             tmk_vmm => TmkVmmOutput,
